+++ packages/webos/platform/platform.js
@@ -1,3 +1,5 @@
+import platform from '../../core/platform';
+
 /**
  * Platform identification of webOS variants
  * @readonly
@@ -6,11 +8,9 @@
  * @property {?boolean} watch - Set true for LG webOS SmartWatch
  * @property {?boolean} open - Set true for Open webOS
  * @property {?boolean} legacy - Set true for legacy webOS (Palm and HP hardware)
- * @property {?boolean} unknown - Set true for any unknown system
 */
-const platform = {};
 
-if (typeof window === 'object' && window.PalmSystem) {
+if (platform.webos || (typeof window === 'object' && window.PalmSystem)) {
 	if (window.navigator.userAgent.indexOf('SmartWatch') > -1) {
 		platform.watch = true;
 	} else if ((window.navigator.userAgent.indexOf('SmartTV') > -1) || (window.navigator.userAgent.indexOf('Large Screen') > -1)) {
@@ -33,9 +33,6 @@ if (typeof window === 'object' && window.PalmSystem) {
 		window.Mojo = window.Mojo || {relaunch: function () {}};
 		if (window.PalmSystem.stageReady) window.PalmSystem.stageReady();
 	}
-} else {
-	platform.unknown = true;
 }
-
 export default platform;
 export {platform};