+++ packages/webos/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact webos module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+- dependency on `@enact/core`
+
+### Changed
+- `webos/platform` to not set `platform.unknown` to `true` (it is handled in `@enact/core/platform`)
+
 ## [1.1.0] - 2017-04-21
 
 No significant changes.